refactor: add hooks & utils to Budget page logic
